Move task for configuration to separate file, determine which config to take based on env parameter
