+++ .travis.yml
@@ -12,4 +12,4 @@ install:
     - node_modules/.bin/bower install --config.interactive=false
 
 script:
-    - node_modules/.bin/gulp test
+    - node_modules/.bin/gulp --env test test
